Try TImeout

+++ app/Http/Controllers/Assignments/UserController.php
@@ -55,7 +55,40 @@ class UserController extends Controller
 
 
 
-
+function ExecWaitTimeout($cmd, $timeout=5) {
+ 
+  $descriptorspec = array(
+      0 => array("pipe", "r"),
+      1 => array("pipe", "w"),
+      2 => array("pipe", "w")
+  );
+  $pipes = array();
+ 
+  $timeout += time();
+  $process = proc_open($cmd, $descriptorspec, $pipes);
+  if (!is_resource($process)) {
+    throw new Exception("proc_open failed on: " . $cmd);
+  }
+ 
+  $output = '';
+ 
+  do {
+    $timeleft = $timeout - time();
+    $read = array($pipes[1]);
+    stream_select($read, $write = NULL, $exeptions = NULL, $timeleft, NULL);
+ 
+    if (!empty($read)) {
+      $output .= fread($pipes[1], 8192);
+    }
+  } while (!feof($pipes[1]) && $timeleft > 0);
+ 
+  if ($timeleft <= 0) {
+    proc_terminate($process);
+    throw new Exception("command timeout on: " . $cmd);
+  } else {
+    return $output;
+  }
+}
 
 public function push(Request $request)
 	{
@@ -127,48 +160,21 @@ public function push(Request $request)
 
         	$destinationPath2 = storage_path() . '//assignments//'.$asn->id.'//user_upload//';
         	$result = shell_exec('javac -d '.$destinationPath.' '.$final);
+          //  dd($result);
+       //     return view('assignments.infinity');
+            $checkpath = $destinationPath2.' '.$filename.'.class';
+            if(!file_exists($checkpath))
+                return view('assignments.compileerror');
+
 			$filename = str_replace(".java","",$filename);
 			$inputpath = storage_path() . '//assignments//'.$request->idc.'//input//';
 			//dd($asn->finput);
 			$injection = 'java -cp '.$destinationPath2.' '.$filename.' < '.$asn->finput.' > '.$destinationPath2.$filename.'.txt';
+            $result = shell_exec('/bin/bash storage_path() $injection');
+            //shell_exec($injection)
 			//dd($injection);
             //INJECTIN ZONE
-            function ExecWaitTimeout($cmd, $timeout=5) {
- 
-  $descriptorspec = array(
-      0 => array("pipe", "r"),
-      1 => array("pipe", "w"),
-      2 => array("pipe", "w")
-  );
-  $pipes = array();
- 
-  $timeout += time();
-  $process = proc_open($cmd, $descriptorspec, $pipes);
-  if (!is_resource($process)) {
-    throw new Exception("proc_open failed on: " . $cmd);
-  }
- 
-  $output = '';
- 
-  do {
-    $timeleft = $timeout - time();
-    $read = array($pipes[1]);
-    stream_select($read, $write = NULL, $exeptions = NULL, $timeleft, NULL);
- 
-    if (!empty($read)) {
-      $output .= fread($pipes[1], 8192);
-    }
-  } while (!feof($pipes[1]) && $timeleft > 0);
- 
-  if ($timeleft <= 0) {
-    proc_terminate($process);
-    throw new Exception("command timeout on: " . $cmd);
-  } else {
-    return $output;
-  }
-}
-            $ticktock = ExecWaitTimeout($injection);
-            dd($ticktock);
+
 
 /*
 
@@ -302,80 +308,10 @@ public function push(Request $request)
                 */
 
             
-            
-            echo("<br>DONE");
-            exit();
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
             // INJECTION ZONE
-        	$result = shell_exec($injection);
+        	//$result = shell_exec($injection);
         	//dd($result);
         	//$injector = storage_path() . '//assignments//'.$request->idc.'//master//';
         	//dd($result);
@@ -394,6 +330,7 @@ public function push(Request $request)
         	$whatsap = strcmp($restore, $geter);
             if($whatsap == 0)
                 $sum+=$per_asn;
+            dd($whatsap);exit();
 //print_r($whatsap);
             //echo($sum);
         	//dd($whatsap);
